# docs/swagger.yaml
basePath: /
definitions:
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  entities.Anomalia:
    properties:
      anomalia_id:
        type: integer
      descripcion:
        type: string
      estado:
        type: string
      fecha_reporte:
        type: string
      fecha_resolucion:
        type: string
      id_chofer_id:
        type: integer
      punto_id:
        type: integer
      tipo_anomalia:
        type: string
    type: object
  entities.AnomaliaListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.Anomalia'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.AnomaliaMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.AnomaliaResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Anomalia'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.CreateAnomaliaRequest:
    properties:
      descripcion:
        type: string
      estado:
        type: string
      fecha_reporte:
        type: string
      id_chofer_id:
        type: integer
      punto_id:
        type: integer
      tipo_anomalia:
        type: string
    required:
    - descripcion
    - id_chofer_id
    - tipo_anomalia
    type: object
...
  entities.CreateSeguimientoFallaCriticaRequest:
    properties:
      comentario:
        type: string
      falla_id:
        type: integer
    required:
    - comentario
    - falla_id
    type: object
...
  entities.Incidencia:
    properties:
      conductor_id:
        type: integer
      created_at:
        type: string
      descripcion:
        type: string
      eliminado:
        type: boolean
      fecha_reporte:
        type: string
      incidencia_id:
        type: integer
      json_ruta:
        type: string
      punto_recoleccion_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  entities.IncidenciaMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.IncidenciaResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Incidencia'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.ReporteConductor:
    properties:
      camion_id:
        type: integer
      conductor_id:
        type: integer
      created_at:
        type: string
      descripcion:
        type: string
      reporte_id:
        type: integer
      ruta_id:
        type: integer
    type: object
...
  entities.ReporteConductorResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.ReporteConductor'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.SeguimientoFallaCritica:
    properties:
      comentario:
        type: string
      created_at:
        type: string
      falla_id:
        type: integer
      seguimiento_id:
        type: integer
    type: object
  entities.SeguimientoFallaCriticaListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.SeguimientoFallaCritica'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.SeguimientoFallaCriticaResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.SeguimientoFallaCritica'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.UpdateAnomaliaRequest:
    properties:
      descripcion:
        type: string
      estado:
        type: string
      fecha_reporte:
        type: string
      fecha_resolucion:
        type: string
      id_chofer_id:
        type: integer
      punto_id:
        type: integer
      tipo_anomalia:
        type: string
    type: object
...
  entities.UpdateIncidenciaRequest:
    properties:
      conductor_id:
        type: integer
      descripcion:
        type: string
      fecha_reporte:
        type: string
      json_ruta:
        type: string
      punto_recoleccion_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/anomalias/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar anomalías
      tags:
      - Anomalia
    post:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateAnomaliaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear anomalía
      tags:
      - Anomalia
  /api/anomalias/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar anomalía
      tags:
      - Anomalia
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Anomalía por ID
      tags:
      - Anomalia
    put:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateAnomaliaRequest'
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AnomaliaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar anomalía
      tags:
      - Anomalia
...
  /api/incidencias/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.IncidenciaMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar incidencia
      tags:
      - Incidencia
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.IncidenciaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Incidencia por ID
      tags:
      - Incidencia
    put:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateIncidenciaRequest'
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.IncidenciaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar incidencia
      tags:
      - Incidencia
...
  /api/reportes-conductor/fecha:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ReporteConductorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reportes conductor por fecha
      tags:
      - ReporteConductor
...
  /api/seguimientos-falla-critica/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.SeguimientoFallaCriticaListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar seguimientos falla crítica
      tags:
      - SeguimientoFallaCritica
    post:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateSeguimientoFallaCriticaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.SeguimientoFallaCriticaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear seguimiento falla crítica
      tags:
      - SeguimientoFallaCritica
...
securityDefinitions:
  BearerAuth:
    description: 'Escribe ''Bearer '' seguido de tu token JWT. Ejemplo: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"